Vector cast operations: refactor
